requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
frozenlist==1.4.1
idna==3.6
multidict==6.0.4
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.0
slack-bolt==1.18.1
slack-sdk==3.26.1
typing_extensions==4.9.0
yarl==1.9.4
